Split up parts of binary into modules with subcommands in preparation for adding serve subcommand
