+++ vimwiki-cli/src/subcommand/mod.rs
@@ -1,0 +1,5 @@
+mod convert;
+mod serve;
+
+pub use convert::convert;
+pub use serve::serve;
