+++ vimwiki-cli/src/subcommand/serve.rs
@@ -1,0 +1,3 @@
+use crate::{CommonOpt, ServeSubcommand};
+
+pub fn serve(_cmd: ServeSubcommand, _opt: CommonOpt) {}
